ceph-provisioners: Don't pass pod name as id to provisioners
Passing the pod name as an id prevents the provisioner from
deleting the PV when the pod creating the PV and the pod
deleting the PV aren't the same which is the standard case
for long-lived PVs. By not supplying an explicit id the id
is set to the provisioner name which doesn't change.