ceph-mon: don't overwrite constructed map with current online map
I studied the scripts in ceph-container and it is not present there.
Furthermore it led to problems when the monitor wasn't present in
the on-disk map but had already been initialized in the past.
In this case the monitor would commit suicide. Now the monitor is always
present in the initial on-disk map as the map isn't overwritten by the
current online map.