ceph-mon: remove code to handle Ceph versions prior to Nautilus
This code led to problems because ceph mon enable-msgr2 would sometimes
disable msgr2 for some monitors when min_mon_release was 0 (unknown).
min_mon_release = 0 could occur when an election was in progress and
the mon map was in flux.

ceph-with-helm will no longer work with Mimic or Luminous with this
change.