Enable (d)mclock scheduler by default
This also adds a dependencies to OSD pods that ensures that the Ceph configuration is completely applied when the OSD
starts. This is needed for mclock because if mclock is active a disk benchhmark is inititaed during OSD startup.
Without the dependency the OSDs will race with the bootstrap job and some OSDs will get the mclock config and some won't.
This is only relevant if deploying Ceph for the first time. If the OSDs are already up when mclock is enabled then the
OSDs need to restarted by hand.